Add missing add/sub unsigned intrinsic support.
This got a little complicated when I realized that "char" is unsigned by
default, so I had to change some iterators to use "signed char", and once
I did that I couldn't get unsigned types by prepending "unsigned" to the
type name, so I had to add some more iterators.  I did this the expedient
way which is maybe not the best way.